requirements_full.txt
flask
pyautogui
argon2-cffi
flask-cors
python-vlc
yt-dlp
PyOpenSSL
